requirements.txt
Click==7.0
biopython==1.72